pyproject.toml
[project]
name = "feedback-report-generator"
version = "0.1.0"
description = ""
authors = [
    {name = "Subass18", email = "144251609+Subass18@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "reportlab (>=4.4.0,<5.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
